refactor(auth): remove shared type folder
