+++ packages/auth/src/contracts/index.ts
@@ -1,1 +1,1 @@
-export type { BetterAuthSessionResponse } from '../shared/better-auth-session.ts';
+export type { BetterAuthSessionResponse } from './auth-session.ts';
