--- packages/auth/src/shared/auth-error.ts
@@ -1,29 +1,0 @@
-import {
-	defineErrors,
-	extractErrorMessage,
-	type InferErrors,
-} from 'wellcrafted/error';
-
-export const AuthError = defineErrors({
-	InvalidCredentials: () => ({
-		message: 'Invalid email or password.',
-	}),
-	SignInFailed: ({ cause }: { cause: unknown }) => ({
-		message: `Failed to sign in: ${extractErrorMessage(cause)}`,
-		cause,
-	}),
-	SignUpFailed: ({ cause }: { cause: unknown }) => ({
-		message: `Failed to create account: ${extractErrorMessage(cause)}`,
-		cause,
-	}),
-	SocialSignInFailed: ({ cause }: { cause: unknown }) => ({
-		message: `Social sign-in failed: ${extractErrorMessage(cause)}`,
-		cause,
-	}),
-	SignOutFailed: ({ cause }: { cause: unknown }) => ({
-		message: `Failed to sign out: ${extractErrorMessage(cause)}`,
-		cause,
-	}),
-});
-
-export type AuthError = InferErrors<typeof AuthError>;
