--- packages/auth/src/shared/better-auth-session.ts
@@ -1,11 +1,0 @@
-import { EncryptionKeys } from '@epicenter/encryption';
-import type {
-	User as BetterAuthUser,
-	Session as BetterSession,
-} from 'better-auth';
-
-export type BetterAuthSessionResponse = {
-	user: BetterAuthUser;
-	session: BetterSession;
-	encryptionKeys: EncryptionKeys;
-};
