--- packages/auth/src/shared/index.ts
@@ -1,8 +1,0 @@
-export * from './auth-error.ts';
-export {
-	type AuthChangeListener,
-	type AuthIdentity,
-	AuthUser,
-	BearerSession,
-} from './auth-types.ts';
-export type { BetterAuthSessionResponse } from './better-auth-session.ts';
